move save to geojson to helpters
