+++ scripts/build_bus_regions.py
@@ -43,6 +43,7 @@ Description
 
 import logging
 from _helpers import configure_logging
+from _helpers import save_to_geojson
 
 import pypsa
 import os
@@ -53,22 +54,6 @@ from shapely.geometry import Polygon
 from scipy.spatial import Voronoi
 
 logger = logging.getLogger(__name__)
-
-
-def save_to_geojson(df, fn):
-    if os.path.exists(fn):
-        os.unlink(fn)
-    if not isinstance(df, gpd.GeoDataFrame):
-        df = gpd.GeoDataFrame(dict(geometry=df))
-    # if geodataframe is not empty. Save shapes.
-    if df.shape[0] > 0:
-        df = df.reset_index()
-        schema = {**gpd.io.file.infer_schema(df), "geometry": "Unknown"}
-        df.to_file(fn, driver="GeoJSON", schema=schema)
-    # if geodataframe is empty, save empty file. See issue 265.
-    else:
-        with open(fn, "w") as fp:
-            pass
 
 
 def voronoi_partition_pts(points, outline):
